requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
